+++ src/encoder/cabac_path.rs
@@ -2013,6 +2013,7 @@ impl Encoder {
             vui: cfg.vui.clone(),
         });
         let pps_cfg = BaselinePpsConfig {
+            redundant_pic_cnt_present_flag: false,
             pic_scaling_lists: cfg.scaling_matrix.pic_spec(),
             chroma_format_idc: cfg.chroma_format_idc,
             pic_parameter_set_id: 0,